move StateProvider to Home Page
