+++ src/components/Posts/AllPost/AllPost.js
@@ -5,10 +5,10 @@ import Container from '@material-ui/core/Container';
 import cookie from 'react-cookies';
 import { serviceRequest } from '../../../services/serviceRequest';
 import config from '../../../config';
-import { StateProvider } from '../../StateProvider';
 import { ErrorMessage } from '../../ErrorMessage';
 import { MainPost } from '../MainPost';
 import { SubPost } from '../SubPost';
+import { useStateValue } from '../../StateProvider';
 
 const path = '/api/v1/user/getposts?';
 const domain = config.apiDomain;
@@ -24,39 +24,26 @@ function preparePayload(method, headers, params) {
 }
 
 export const AllPost = () => {
-  const [allPostParams] = useState({ page: 1, pageSize: 15, viewall: true });
-  const token = cookie.load('jwtToken');
-  const allPostHeaders = { Authorization: token };
-  const [allPosts, setAllPosts] = useState({ posts: [] });
+  const [{ posts, page }, dispatch] = useStateValue();
   const [isError, setIsError] = useState(false);
   const [errorMsg, setErrorMsg] = useState('Authorization Error');
 
-  const subPosts = allPosts.posts.slice(1);
-
-  const initialState = {
-    page: { offset: 0, limit: 1 },
-  };
-
-  const reducer = (state, action) => {
-    switch (action.type) {
-      case 'changePage':
-        return {
-          ...state,
-          page: action.newPage,
-        };
-
-      default:
-        return state;
-    }
-  };
+  const token = cookie.load('jwtToken');
+  const allPostHeaders = { Authorization: token };
+  const subPosts = posts.slice(1);
 
   useEffect(() => {
     async function fetchAllPosts() {
-      const requestPayload = preparePayload('get', allPostHeaders, allPostParams);
+      const requestPayload = preparePayload('get', allPostHeaders,
+        { page: page.currentPage, pageSize: page.pageSize, viewall: true });
       const response = await serviceRequest(requestPayload);
       try {
         if (response.status && response.status === 'success') {
-          setAllPosts(response.data);
+          const { data } = response;
+          dispatch({
+            type: 'changePosts',
+            posts: data.posts,
+          });
         } else if (response.status && response.status === 'error') {
           setErrorMsg('Authorization Error');
           setIsError(true);
@@ -69,17 +56,18 @@ export const AllPost = () => {
       }
     }
     fetchAllPosts();
-  }, [initialState.page.offset, initialState.page.limit]);
+    // eslint-disable-next-line
+  }, [posts.currentPage]);
 
   return (
-    <StateProvider initialState={initialState} reducer={reducer}>
+    <>
       {isError ? (<ErrorMessage message={errorMsg} styles={{ color: 'red' }} />)
         : (
           <>
             <CssBaseline />
             <Container maxWidth="lg">
               <main>
-                <MainPost mainPost={allPosts.posts[0]} />
+                <MainPost mainPost={posts[0]} />
                 <Grid container spacing={4}>
                   {subPosts.map((post) => (
                     <SubPost post={post} key={post.posterID + post.title} />
@@ -89,6 +77,6 @@ export const AllPost = () => {
             </Container>
           </>
         )}
-    </StateProvider>
+    </>
   );
 };
