+++ src/components/StateProvider/StateProvider.js
@@ -1,6 +1,7 @@
+/* eslint-disable react/forbid-prop-types */
 import React, { createContext, useContext, useReducer } from 'react';
 import {
-  func, objectOf, object, oneOfType, node,
+  func, objectOf, object, oneOfType, node, any,
 } from 'prop-types';
 
 export const StateContext = createContext();
@@ -15,7 +16,7 @@ export const useStateValue = () => useContext(StateContext);
 
 StateProvider.propTypes = {
   reducer: func.isRequired,
-  initialState: objectOf(object).isRequired,
+  initialState: any.isRequired,
   children: oneOfType([
     node,
     func,
